+++ stdlib/number/round_test.go
@@ -1,6 +1,7 @@
 package numberfunc_test
 
 import (
+	"math/big"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-framework/attr"
@@ -38,6 +39,15 @@ func TestRoundFunction(test *testing.T) {
 			},
 			Expected: function.RunResponse{
 				Result: function.NewResultData(types.Int64Value(2)),
+			},
+		},
+		"beyond-upper-limit": {
+			Request: function.RunRequest{
+				Arguments: function.NewArgumentsData([]attr.Value{types.NumberValue(func() *big.Float { f := new(big.Float); f.SetString("1e+310"); return f }())}),
+			},
+			Expected: function.RunResponse{
+				Result: resultData,
+				Error:  function.NewFuncError("round: input number is beyond the limits of float64 for rounding"),
 			},
 		},
 	}